split out tremolo and vibrato logic into separate files
